+++ cmd/plugins.go
@@ -14,6 +14,7 @@ import (
 	_ "github.com/tsarna/vinculum/servers/vws"
 	_ "github.com/tsarna/vinculum/servers/websocket"
 	_ "github.com/tsarna/vinculum/triggers/after"
+	_ "github.com/tsarna/vinculum/triggers/at"
 	_ "github.com/tsarna/vinculum/triggers/cron"
 	_ "github.com/tsarna/vinculum/triggers/interval"
 	_ "github.com/tsarna/vinculum/triggers/once"