+++ triggers/at/testplugins_test.go
@@ -1,0 +1,6 @@
+package at_test
+
+// Import functions package to trigger init() registrations during testing,
+// making time-cty-funcs functions (now(), timeadd(), duration(), etc.)
+// available in eval contexts.
+import _ "github.com/tsarna/vinculum/functions"